requirements.txt
flask
flask-socketio
eventlet
bcrypt
gunicorn
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
google-generativeai
flask-caching